refactor(cowork): split voice input modules
