+++ src/main/ipcHandlers/asr/index.ts
@@ -1,0 +1,2 @@
+export type { AsrHandlerDeps } from './handlers';
+export { registerAsrIpcHandlers } from './handlers';
