+++ src/renderer/services/voiceInput/constants.ts
@@ -1,0 +1,4 @@
+export const VOICE_INPUT_TARGET_SAMPLE_RATE = 16000;
+export const VOICE_INPUT_MAX_RECORDING_MS = 60_000;
+
+export const buildVoiceInputFileName = (): string => `voice-input-${Date.now()}.wav`;
